add fake event preprocessor
