+++ src/test/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/JniEventPreprocessorTest.kt
@@ -22,7 +22,7 @@ import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 import org.wfanet.panelmatch.client.PreprocessEventsRequestKt.unprocessedEvent
 import org.wfanet.panelmatch.client.copy
-import org.wfanet.panelmatch.client.eventpreprocessing.testing.EventPreprocessorTest
+import org.wfanet.panelmatch.client.eventpreprocessing.testing.AbstractEventPreprocessorTest
 import org.wfanet.panelmatch.client.preprocessEventsRequest
 import org.wfanet.panelmatch.common.JniException
 import org.wfanet.panelmatch.common.compression.CompressionParametersKt.noCompression
@@ -41,7 +41,7 @@ private val REQUEST = preprocessEventsRequest {
 }
 
 @RunWith(JUnit4::class)
-class JniEventPreprocessorTest : EventPreprocessorTest() {
+class JniEventPreprocessorTest : AbstractEventPreprocessorTest() {
   override val eventPreprocessor: EventPreprocessor = JniEventPreprocessor()
 
   @Test
