+++ src/test/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/testing/FakeEncryptEventsDoFnTest.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.eventpreprocessing
+package org.wfanet.panelmatch.client.eventpreprocessing.testing
 
 import com.google.protobuf.ByteString
 import com.google.protobuf.kotlin.toByteStringUtf8
@@ -22,15 +22,15 @@ import org.apache.beam.sdk.coders.ListCoder
 import org.apache.beam.sdk.extensions.protobuf.ByteStringCoder
 import org.apache.beam.sdk.transforms.DoFn
 import org.apache.beam.sdk.transforms.ParDo
-import org.apache.beam.sdk.transforms.SerializableFunction
 import org.apache.beam.sdk.values.KV
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
-import org.wfanet.panelmatch.client.PreprocessEventsRequest
-import org.wfanet.panelmatch.client.PreprocessEventsResponse
-import org.wfanet.panelmatch.client.PreprocessEventsResponseKt.processedEvent
-import org.wfanet.panelmatch.client.preprocessEventsResponse
+import org.wfanet.panelmatch.client.eventpreprocessing.EncryptEventsDoFn
+import org.wfanet.panelmatch.client.eventpreprocessing.EventPreprocessorFn
+import org.wfanet.panelmatch.client.eventpreprocessing.HardCodedDeterministicCommutativeCipherKeyProvider
+import org.wfanet.panelmatch.client.eventpreprocessing.HardCodedHkdfPepperProvider
+import org.wfanet.panelmatch.client.eventpreprocessing.HardCodedIdentifierHashPepperProvider
 import org.wfanet.panelmatch.common.beam.kvOf
 import org.wfanet.panelmatch.common.beam.testing.BeamTestBase
 import org.wfanet.panelmatch.common.beam.testing.assertThat
@@ -43,7 +43,7 @@ private const val EXPECTED_SUFFIX = "$IDENTIFIER_HASH_PEPPER$HKDF_HASH_PEPPER$CR
 
 /** Unit tests for [EncryptEventsDoFn]. */
 @RunWith(JUnit4::class)
-class EncryptEventsDoFnTest : BeamTestBase() {
+class FakeEncryptEventsDoFnTest : BeamTestBase() {
   @Suppress("NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS")
   private val coder: Coder<MutableList<KV<ByteString, ByteString>>> =
     ListCoder.of(KvCoder.of(ByteStringCoder.of(), ByteStringCoder.of()))
@@ -59,7 +59,7 @@ class EncryptEventsDoFnTest : BeamTestBase() {
 
     val doFn: DoFn<MutableList<KV<ByteString, ByteString>>, KV<Long, ByteString>> =
       EncryptEventsDoFn(
-        FakeEncryptEvents,
+        EventPreprocessorFn(FakeEventPreprocessor()),
         HardCodedIdentifierHashPepperProvider(IDENTIFIER_HASH_PEPPER.toByteStringUtf8()),
         HardCodedHkdfPepperProvider(HKDF_HASH_PEPPER.toByteStringUtf8()),
         HardCodedDeterministicCommutativeCipherKeyProvider(CRYPTO_KEY.toByteStringUtf8()),
@@ -81,20 +81,4 @@ private fun outputOf(key: Long, value: String): KV<Long, ByteString> {
 
 private fun expectedOutputOf(key: Long, value: String): KV<Long, ByteString> {
   return outputOf(key, value + EXPECTED_SUFFIX)
-}
-
-private object FakeEncryptEvents :
-  SerializableFunction<PreprocessEventsRequest, PreprocessEventsResponse> {
-  override fun apply(request: PreprocessEventsRequest): PreprocessEventsResponse {
-    val suffix = with(request) { identifierHashPepper.concat(hkdfPepper).concat(cryptoKey) }
-    return preprocessEventsResponse {
-      for (events in request.unprocessedEventsList) {
-        processedEvents +=
-          processedEvent {
-            encryptedId = events.id.toStringUtf8().toLong() + 1
-            encryptedData = events.data.concat(suffix)
-          }
-      }
-    }
-  }
 }
